[Pathway 3-2] 9. Extract the NavHost into RallyNavHost
